Initial openvpn/shapeshifter tcp/udp configuration.
In order for openvpn to bring up and modify the tun interfaces it needs, it must
have access to /dev/net and have the NET_ADMIN capability. To achieve this, we
mount /dev/net into the container, and add capbility NET_ADMIN to the openvpn
container.

Install suitable openvpn DNS records (the formatting needs to be such that
zonetool can parse it).

Setup firewall rules to properly forward openvpn traffic for clients.