Allow running the openvpn service on port 443
This change allows us to share port 443 between OpenVPN and NGINX,
when the openvpn hosts overlap with the frontend hosts. The change
happens automatically because we inject a role before the 'frontend'
role, so we can detect the overlap and tell NGINX to run on a
different port. We then use 'sslh' to de-multiplex the SSL protocols
and route traffic to either openvpn or nginx.