Update README.md to clarify more things.
Specify what machines will work, the minimal configuration of the system and
detail what float will manage, and what needs to be managed externally.

Detail what is needed in terms of DNS and how that will be managed.

Suggest using an agent so you aren't prompted for the password.

Clarify what should be put into the config.yaml sections and provide a pointer
to additional documentation for setting up U2F and OTP tokens.