Handle case where gateway_address6 is not set.
If a gateway does not have ipv6, it should not be setting a gateway_address6 in
the hosts.yml. If it does, then it will get rendered in the eip-service.json and
clients will attempt to contact it.

If you simply removed that configuration from the gateway in hosts.yml, then it
would get rendered in the eip-service.json as `gateway_address6: null` which
would also cause clients to attempt to contact that malformed address.

This commit will strip out any nulls (but not False) from the eip-service.json
and the provider.json.