Add additional listening ports (#16).
This will listen on port 1194 and port 53 and redirect traffic to
openvpn (running now on port 80). It will also advertise those ports in the
`eip-service.json`.

In order to accomplish this, we need to have the knot-resolver listen explicitly
on the openvpn ip, instead of being greedy with its ports.