Fix Go dependency install instructions in README
Prior to Go 1.17 the "go get" command was overloaded to install library
dependencies and applications. As of Go 1.17 installing applications
with it results in a warning, and as of Go 1.18 it will no longer be
used to install applications.

This patch updates the readme to use "go install" instead.

TL;DR — "go get" updates module dependencies, "go install" builds
        binaries and sticks them in $GOBIN.

Signed-off-by: Sam Whited <sam@samwhited.com>