Enable an optional custom vpnweb domain.
For transition from the old platform, it can be useful to have the vpnweb
respond to another domain. For example, calyx responds now on api.calyx.net,
however the vpnweb with lilypad will appear on api.vpn.calyx.net. With this
change it is possible to configure `custom_vpn_web_domains: [api.calyx.net]` to
have it available there as well.