Abstract out simplevpn into its own role.
It was possible for the `provider` service to not get an updated provider.json
when it is generated if it is scheduled on a different backend than `vpnweb`,
because it didn't have its own role to apply that template. This fixes that
missing piece for the `provider` service.

However, because the provider role and the vpnweb both need to be able to
reference the templates that are rendered by simplevpn, this abstracts that into
its own role so both can use it.